backend/requirements.txt
# Core browser automation
browser-use

# LangChain integrations for different LLM providers
langchain-openai
langchain-anthropic
langchain-google-genai

# Web framework (for API if needed)
fastapi
uvicorn

# Environment variables
python-dotenv

# PDF handling
PyPDF2>=3.0.0
fillpdf>=0.7.0

# HTTP client for PDF downloads
aiohttp>=3.9.0
